requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pandas>=2.1.0
requests>=2.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
